feat(naming): add ENRENameFile type and implementation
