+++ packages/enre-naming/src/utils/xml-builder.ts
@@ -1,0 +1,11 @@
+import allowedTags from '../xml/allowed-tags';
+
+export default (tag: typeof allowedTags[number], props: Record<string, string>) => {
+  let tmp = '<' + tag;
+
+  tmp = Object.keys(props).reduce((prev, curr) =>
+      prev + ` ${curr}="${props[curr]}"`
+    , tmp);
+
+  return tmp + '>';
+};
