+++ packages/enre-naming/src/xml/allowed-tags.ts
@@ -1,0 +1,5 @@
+export default [
+  'File',
+  'Anonymous',
+  'Modified',
+] as const;
