+++ packages/enre-naming/tests/index.test.js
@@ -5,6 +5,48 @@ describe('Build ENRE name', () => {
     expect(buildENREName('foo')).toMatchSnapshot();
   });
 
+  describe('with file payload', () => {
+    it('js', () => {
+      expect(buildENREName({base: 'file0', ext: 'js'})).toMatchSnapshot();
+    });
+
+    it('mjs', () => {
+      expect(buildENREName({base: 'file0', ext: 'mjs'})).toMatchSnapshot();
+    });
+
+    it('cjs', () => {
+      expect(buildENREName({base: 'file0', ext: 'cjs'})).toMatchSnapshot();
+    });
+
+    it('jsx', () => {
+      expect(buildENREName({base: 'file0', ext: 'jsx'})).toMatchSnapshot();
+    });
+
+    it('ts', () => {
+      expect(buildENREName({base: 'file0', ext: 'ts'})).toMatchSnapshot();
+    });
+
+    it('d.ts', () => {
+      expect(buildENREName({base: 'file0', ext: 'd.ts'})).toMatchSnapshot();
+    });
+
+    it('mts', () => {
+      expect(buildENREName({base: 'file0', ext: 'mts'})).toMatchSnapshot();
+    });
+
+    it('cts', () => {
+      expect(buildENREName({base: 'file0', ext: 'cts'})).toMatchSnapshot();
+    });
+
+    it('tsx', () => {
+      expect(buildENREName({base: 'file0', ext: 'tsx'})).toMatchSnapshot();
+    });
+
+    it('json', () => {
+      expect(buildENREName({base: 'file0', ext: 'json'})).toMatchSnapshot();
+    });
+  });
+
   describe('with anonymous payload', () => {
     it('as Function', () => {
       expect(buildENREName({as: 'Function'})).toMatchSnapshot();
